its almost done
the general code structure is better and less error prone and it doesnt spam suppliers everywhere.
the goal is also sometimes null for some reason

TODO tmr:
dont immediately score when NBA, have a separate button for scoring during NBA and be able to move left and right.
dont be able to move left and right during AQUA.